From-SVN: r227693

+++ gcc/testsuite/c-c++-common/torture/vector-shift1.c
@@ -2,7 +2,7 @@
 #define vector __attribute__((vector_size(8*sizeof(short))))
 
 int main (int argc, char *argv[]) {
-  vector short v0 = {argc,2,3,4,5,6,7};
+  vector short v0 = {(short)argc,2,3,4,5,6,7};
   vector short v1 = {2,2,2,2,2,2,2};
   vector short r1,r2,r3,r4;
   int i = 8;
